refactor: move entities to the domain package
